gui.page: make `PageWidget` page size a pair of integers.

+++ gui/page.py
@@ -56,7 +56,7 @@ class PageWidget(wx.Panel):
 				page_job = page.decode(wait = False)
 				page_text = page.text
 				dpi = float(page_job.dpi)
-				page_size = tuple(t * 100.0 / dpi for t in (page_job.width, page_job.height))
+				page_size = tuple(int(t * 100.0 / dpi) for t in (page_job.width, page_job.height))
 				self.SetSize(page_size)
 				self.SetBestFittingSize(page_size)
 				self.GetParent().Layout()
